// src/types.tsx
export type User = {
  _id?: string;
  userId?: string;
  firstName: string;
  lastName: string;
  wallet?: Wallet;
  email: string;
  phone: string;
  password: string;
  birthday: string;
  home_address?: string;
  office_address?: string;
  primary_interest?: string;
  current_challenges?: string;
  is_demo: number;
  preferred_date?: string;
  last_login: Date;
  date_joined: Date;
  suspended?: boolean;
  del_falg: number;
  verified: boolean;
  salt: string;
  role: any;
  otp: number;
  otpExpiryTime: Date;
  profile_pic?: string;

  title?: "Mr." | "Ms." | "Mrs." | "Dr." | "Prof.";
  gender?: "Male" | "Female" | "Other";
  religion?: string;
  place_of_birth?: string;
  current_parish?: string;
  national_id_number?: string;
  national_id_photo?: string;
  home_location_map?: string;
  district_of_birth?: string;
  parish_of_birth?: string;
  village_of_birth?: string;
  marital_status?: "Single" | "Married" | "Divorced" | "Widowed";
  occupation?: string;
  job_title?: string;
  next_of_kin_name?: string;
  next_of_kin_phone?: string;
  next_of_kin_email?: string;
  next_of_kin_national_id?: string;
  monthly_income_level?:
    | "Less than UGX 1,000,000"
    | "UGX 1,000,000 - 5,000,000"
    | "UGX 5,000,000 - 15,000,000"
    | "Above UGX 15,000,000";
  bank_name?: string;
  bank_account_number?: string;
  bank_mobile_account?: string;
  bank_email?: string;
  highest_education_level?:
    | "Secondary (Ordinary Level)"
    | "Secondary (Advanced Level)"
    | "Tertiary"
    | "University"
    | "Other (Specify)";
  employment_status?: "Employed" | "Self-employed" | "Unemployed" | "Retired";
  current_work_address?: string;
  employer_name?: string;
  current_salary?: string;
  side_hustle_income?: string;
  securityQuestions: {
    question: string;
    answer: string;
  }[];
  is_complete: boolean;
  socketId?: string;
  isActive?: boolean;
  createdAt?: Date;
  id: string;
  active_status?: "online" | "offline" | "do not disturb" | "idle";
};
// ...
export interface Transaction {
  amount?: number | string;
  createdAt?: Date;
  date?: Date;
  user?: User;
  wallet?: string;
  _id?: string;
  type?: string;
}

export interface BF {
  fundName: string;
  fundDetails: string;
  accountType: "bank" | "mobile" | "wallet";
  accountInfo: string;
  walletAddress?: string;
  groupId?: string;
  group?: GroupTypes;
  createdBy: User;
  _id?: string;
  role?: string;
  wallet?: Wallet;
  contributions?: Contribution[];
  createdAt?: Date;
  members?: User[];
}

export interface Wallet {
  walletAddress: string;
  balance: number;
  transactionHistory: Transaction[];
}